+++ src/constants/emojis.ts
@@ -7,4 +7,10 @@ export const POLL_REACTIONS = [
   '9️⃣', '🔟'
 ];
 
+export const ISSUE_TYPE_ICONS = {
+  bug: '🐞',
+  feature: '💡',
+  unknown: '❔'
+};
+
 /* eslint-enable @typescript-eslint/naming-convention */