+++ tests/suites/commands/issue.ts
@@ -1,9 +1,8 @@
 import { Client, Command } from '../../../src/structures';
 import { createGuildMessage } from '../../utils/message';
 
-export const ping = (client: Client) => describe('Ping', () => {
+export const about = (client: Client) => describe('About', () => {
   const message = createGuildMessage(client);
-  const spies: { [key: string]: jest.SpyInstance } = {};
   let command: Command; // eslint-disable-line @typescript-eslint/init-declarations
   let args: any = {};   // eslint-disable-line @typescript-eslint/init-declarations
 
@@ -15,16 +14,12 @@ export const ping = (client: Client) => describe('Ping', () => {
 
     command = <Command> client.commands.modules.get(name!);
     if (command && rest.length) args = await command.parse(message, rest);
-
-    spies.embed = jest.spyOn(command, 'embed');
   };
 
-  it('should reply with Pong!', async () => {
-    await setup('!ping');
-    const sent = await command.exec(message, args);
-    expect(spies.embed).toBeCalledTimes(2);
-    expect(sent.embeds).toHaveLength(1);
-    expect(sent.embeds[0].title).toEqual('COMMAND_PING_RESPONSE_TITLE');
+  it('should reply with information', async () => {
+    await setup('!about');
+    await command.exec(message, args);
+    expect(message.channel.send).toBeCalledTimes(1);
   });
 
 });